rename module from fallback to data
